refactor(notice): 기존 posts폴더에 있던 공고 목록을 랜딩페이지로 변경
